Update date 24.03.17

+++ chapter4/1.SRP/src/main/java/com/greensnow25/CalculatorAction.java
@@ -14,5 +14,5 @@ public interface CalculatorAction {
      * @param two number.
      * @return result.
      */
-    int makeAction(int one, int two);
+    double makeAction(double one, double two);
 }
